Moving AESEncryptionResult into its own file and redoing imports.
